feat(UI-1031): add google analytics - refactor
