+++ src/hooks/index.ts
@@ -2,7 +2,6 @@ export { useFetchTrigger } from "@hooks/triggers/useFetchTrigger";
 export { useConnectionForm } from "@hooks/useConnectionForm";
 export { useCreateProjectFromTemplate } from "@hooks/useCreateProjectFromTemplate";
 export { useFileOperations } from "@hooks/useFileOperations";
-export { useGoogleAnalytics } from "@hooks/useGoogleAnalytics";
 export { useInterval } from "@hooks/useInterval";
 export { useResize } from "@hooks/useResize";
 export { useSort } from "@hooks/useSort";
